fix(audit): pass 6 — int-type the shared spurious-pair constant
kSpuriousMinimizePairMs was qint64, narrowing at the autotile debounce
pin (QTimer wants int) and mixing widths in the gap comparison. Both
consumers do sub-second millisecond arithmetic, so int is the right
type end to end.

+++ kwin-effect/plasmazoneseffect.h
@@ -103,7 +103,7 @@ static_assert(
 // (autotilehandler/signals.cpp) and the minimize shader-event
 // spurious-pair cancel (plasmazoneseffect/daemon_apply.cpp,
 // slotWindowMinimizedChanged). Shared here so the two can never desync.
-inline constexpr qint64 kSpuriousMinimizePairMs = 75;
+inline constexpr int kSpuriousMinimizePairMs = 75;
 
 // Forward declarations for helper classes
 class AutotileHandler;
